Introduce support of std::hash for vk-structures.
